+++ src/containerapp/azext_containerapp/_params.py
@@ -48,6 +48,7 @@ def load_arguments(self, _):
     with self.argument_context('containerapp create', arg_group='Runtime', is_preview=True) as c:
         c.argument('runtime', arg_type=get_enum_type(['generic', 'java']), help='The runtime of the container app.', is_preview=True)
         c.argument('enable_java_metrics', arg_type=get_three_state_flag(), help='Boolean indicating whether to enable Java metrics for the app. Only applicable for Java runtime.', is_preview=True)
+        c.argument('enable_java_agent', arg_type=get_three_state_flag(), help='Boolean indicating whether to enable Java agent for the app. Only applicable for Java runtime.', is_preview=True)
 
     # Source and Artifact
     with self.argument_context('containerapp update') as c:
@@ -67,6 +68,7 @@ def load_arguments(self, _):
     with self.argument_context('containerapp update', arg_group='Runtime', is_preview=True) as c:
         c.argument('runtime', arg_type=get_enum_type(['generic', 'java']), help='The runtime of the container app.', is_preview=True)
         c.argument('enable_java_metrics', arg_type=get_three_state_flag(), help='Boolean indicating whether to enable Java metrics for the app. Only applicable for Java runtime.', is_preview=True)
+        c.argument('enable_java_agent', arg_type=get_three_state_flag(), help='Boolean indicating whether to enable Java agent for the app. Only applicable for Java runtime.', is_preview=True)
 
     with self.argument_context('containerapp env', arg_group='Virtual Network') as c:
         c.argument('infrastructure_resource_group', options_list=['--infrastructure-resource-group', '-i'], help='Name for resource group that will contain infrastructure resources. If not provided, a resource group name will be generated.', is_preview=True)
@@ -347,3 +349,16 @@ def load_arguments(self, _):
 
     with self.argument_context('containerapp env', arg_group='Peer Traffic Configuration') as c:
         c.argument('p2p_encryption_enabled', arg_type=get_three_state_flag(), options_list=['--enable-peer-to-peer-encryption'], is_preview=True, help='Boolean indicating whether the peer-to-peer traffic encryption is enabled for the environment.')
+
+    with self.argument_context('containerapp java logger set') as c:
+        c.argument('logger_name', help="The existing logger name to update or new logger name.")
+        c.argument('logger_level', arg_type=get_enum_type(["off", "error", "info", "debug", "trace", "warn"]), help="Set the log level for the specific logger name. Allowed values: off, error, info, debug, trace, warn.")
+
+    with self.argument_context('containerapp java logger delete') as c:
+        c.argument('logger_name', help="The existing logger name to be deleted.")
+        c.argument('all', help="The flag to indicate delete all logger settings.", action="store_true")
+
+    with self.argument_context('containerapp java logger show') as c:
+        c.argument('logger_name', help="The existing logger name to display.")
+        c.argument('all', help="The flag to indicate display all logger settings.", action="store_true")
+