+++ src/containerapp/azext_containerapp/commands.py
@@ -204,3 +204,8 @@ def load_command_table(self, args):
         g.custom_command('update', 'update_eureka_server_for_spring', supports_no_wait=True)
         g.custom_show_command('show', 'show_eureka_server_for_spring')
         g.custom_command('delete', 'delete_eureka_server_for_spring', confirmation=True, supports_no_wait=True)
+
+    with self.command_group('containerapp java logger') as g:
+        g.custom_command('set', 'create_or_update_java_logger', is_preview=True)
+        g.custom_command('delete', 'delete_java_logger', is_preview=True)
+        g.custom_command('show', 'show_java_logger', is_preview=True)